storybookのモバイル設定やっと分かった
