+++ .storybook/main.ts
@@ -1,14 +1,13 @@
 import type { StorybookConfig } from "@storybook/react-vite";
 
 const config: StorybookConfig = {
-  stories: ["../app/**/*.mdx", "../app/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
+  stories: ["../app/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
   addons: [
     "@storybook/addon-onboarding",
     "@chromatic-com/storybook",
     "@storybook/addon-vitest",
     "storybook-addon-remix-react-router",
     "@storybook/addon-themes",
-    "@storybook/addon-links",
     "@storybook/addon-docs",
   ],
   framework: {
