+++ package.json
@@ -32,6 +32,7 @@
     "@radix-ui/react-tooltip": "^1.2.7",
     "@react-router/node": "^7.4.0",
     "@react-router/serve": "^7.4.0",
+    "@storybook/builder-vite": "^9.0.8",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "^1.1.1",
@@ -55,11 +56,11 @@
     "@netlify/vite-plugin-react-router": "^1.0.0",
     "@react-router/dev": "^7.4.0",
     "@shadcn/ui": "^0.0.4",
-    "@storybook/addon-docs": "^9.0.6",
-    "@storybook/addon-onboarding": "^9.0.6",
-    "@storybook/addon-themes": "^9.0.6",
-    "@storybook/addon-vitest": "^9.0.6",
-    "@storybook/react-vite": "^9.0.6",
+    "@storybook/addon-docs": "^9.0.8",
+    "@storybook/addon-onboarding": "^9.0.8",
+    "@storybook/addon-themes": "^9.0.8",
+    "@storybook/addon-vitest": "^9.0.8",
+    "@storybook/react-vite": "^9.0.8",
     "@tailwindcss/postcss": "^4.1.7",
     "@tailwindcss/vite": "^4.0.0",
     "@testing-library/jest-dom": "^6.6.3",
@@ -80,7 +81,7 @@
     "orval": "^7.8.0",
     "playwright": "^1.51.1",
     "react-router-devtools": "^1.1.0",
-    "storybook": "^9.0.6",
+    "storybook": "^9.0.8",
     "storybook-addon-remix-react-router": "^5.0.0",
     "tailwindcss": "^4.1.8",
     "typescript": "^5.7.2",
